requirements.txt
# Database Dependencies
psycopg[binary]==3.3.3

# Environment Management
python-dotenv==1.2.2

# Development Dependencies (Optional)
pytest==7.4.2
pytest-cov==4.1.0